requirements.txt
fastapi[all]
motor[srv]
python-dotenv
uvicorn[standard]
